refactor: cleanup keyvalue store packages
