+++ src/main/java/com/avalon/avalonchat/infra/redis/RedisPhoneNumberAuthCodeStore.java
@@ -5,15 +5,15 @@ import java.time.Duration;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.core.ValueOperations;
 
-import com.avalon.avalonchat.core.user.keyvalue.AuthCodeValue;
-import com.avalon.avalonchat.core.user.keyvalue.KeyAuthCodeValueStore;
-import com.avalon.avalonchat.core.user.keyvalue.PhoneNumberKey;
+import com.avalon.avalonchat.core.user.application.PhoneNumberAuthCodeStore;
+import com.avalon.avalonchat.core.user.application.keyvalue.AuthCodeValue;
+import com.avalon.avalonchat.core.user.application.keyvalue.PhoneNumberKey;
 
-public class RedisPhoneNumberKeyAuthCodeValueStore implements KeyAuthCodeValueStore<PhoneNumberKey> {
+public class RedisPhoneNumberAuthCodeStore implements PhoneNumberAuthCodeStore {
 
 	private final ValueOperations<PhoneNumberKey, AuthCodeValue> phoneNumberAuthOperations;
 
-	public RedisPhoneNumberKeyAuthCodeValueStore(
+	public RedisPhoneNumberAuthCodeStore(
 		RedisTemplate<PhoneNumberKey, AuthCodeValue> phoneNumberAuthRedisTemplate
 	) {
 		this.phoneNumberAuthOperations = phoneNumberAuthRedisTemplate.opsForValue();
